requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pgrep==2019.3.22
public==2019.3.22
PyYAML==5.1
Werkzeug==0.15.2
